Add supplier and purchase stock inward workflows
- Add supplier master entity, repository, service, and REST APIs
- Link products to suppliers through supplier_id while keeping supplierName for compatibility
- Add purchase bill, purchase bill item, and purchase payment entities
- Add purchase APIs for creating and fetching supplier purchase bills
- Automatically increase product stock when purchase bills are saved
- Update product cost price and optional selling price from purchase entries
- Add purchase bill paid/due tracking
- Add purchase cancellation flow with stock reversal
- Track purchase status, cancel reason, and cancelled timestamp
- Treat legacy purchase bills with null status as ACTIVE
- Add integration tests for supplier-linked products, purchase creation, and purchase cancellation

+++ src/main/java/com/pahal/billingApp/entity/BillItem.java
@@ -7,15 +7,20 @@ import lombok.Setter;
 import lombok.ToString;
 
 /**
- * This is a "child" entity. It records which product was sold, at what price, and in what quantity at the time of the sale.
+ * This is a "child" entity. It records which product was sold, at what price,
+ * and in what quantity at the time of the sale.
  *
- * Tip: We store the price here separately because if you change the product price tomorrow, the old bills should still show the price the customer actually paid.
+ * Tip: We store the price here separately because if you change the product
+ * price tomorrow, the old bills should still show the price the customer
+ * actually paid.
  */
 @Entity
 @Table(name = "bill_items")
 @Getter
 @Setter
 @ToString
+// We don't include the Bill reference in equals/hashCode to avoid circular
+// references and potential performance issues.
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 public class BillItem {
     @Id
@@ -24,10 +29,11 @@ public class BillItem {
     private Long id;
 
     private String productName;
-    private Integer quantity;
+    private Double quantity;
 
     /**
-     * Editable selling price per unit at billing time (can differ from product master price).
+     * Editable selling price per unit at billing time (can differ from product
+     * master price).
      */
     private Double unitSellingPrice;
